feat : Made few changes in monster combat pattern. targetting player lower than 30% of maxhp, when there is only one player left monster only uses attackskill, when monster's hp is almost full uses attackskill.

+++ Playable/PlayableC.cs
@@ -7,7 +7,7 @@ public class PlayableC : ScriptableObject
     public Sprite characterImage;
     public int level;
     public float hp; //힐러는 공격력으로 힐을 해서 힐러의 공격력 수치에 주의.
-    public int maxHp;
+    public float maxHp;
     public float mp;
     public int maxMp;
     public float atk;
